requirements.txt
tabulate
mysql-connector-python
python-dotenv